change template and command route
