+++ app/components/build-banner/template.hbs
@@ -162,7 +162,7 @@
     {{#if this.template}}
       <li class="template-info">
         <LinkTo
-          @route="templates.detail"
+          @route="templates.detail.version"
           @models={{array this.template.namespace this.template.name this.template.version}}
         >
           <span class="banner-value">
